refactor(vpc-egress-gateway): rename observer package
(cherry picked from commit 48f890c2ffe1a2b8fe1ff2c947ed72f84f2fdf75)
