+++ pkg/vegobserver/interface_collector.go
@@ -1,4 +1,4 @@
-package vegoobserver
+package vegobserver
 
 import (
 	"encoding/json"
