+++ pkg/vegobserver/observer_test.go
@@ -1,4 +1,4 @@
-package vegoobserver
+package vegobserver
 
 import (
 	"bytes"
