+++ pkg/vegobserver/runner.go
@@ -1,4 +1,4 @@
-package vegoobserver
+package vegobserver
 
 import (
 	"context"
